Add a first POC for a composable activity API
The new API should reduce the complexity of describing a domain story with multiple actors and splits in activity streams.

My special thanks go to Falk(@sippsack) and Tobias(@jopsdev) for an extremely productive and fun ensemble programming session during one of the past CyberLand Open Source Camps.
During the session, the foundations for the new API were laid.

Issue: #11